Giving startOrOptions a default value for IDL conformance.
According to the Web IDL spec, section 2.5.3 (some formatting is mine):

If
- the type of an argument is a dictionary type or a union type that has
a dictionary type as one of its flattened member types,
- and that dictionary type and its ancestors have no required members,
- and the argument is either the final argument or is followed only by
optional arguments,

then the argument must be specified as optional and have a default value
provided.

IIUC, the 'startOrOptions' parameter for Performance.Measure falls into
the category where it needs to be marked optional and given a default
value. Right now it is marked optional but does not have a default
value.

This change adds a default value of an empty dictionary for the
startOrOptions parameter. This should not change any user-visible
behaviour because an omitted paremeter would have been treated as an
empty dictionary in the bindings layer anyways.